// src/constructs/dlz-lake-formation/lake-formation/interfaces.ts
export interface LFTag {
  readonly tagKey: string;
  readonly tagValues: string[];
}

export interface BaseSharedTagProps {
  /**
   * A list of principal identity ARNs (e.g., AWS accounts, IAM roles/users) that the permissions apply to.
   */
  readonly principals: string[];
  /**
   * OPTIONAL - A list of specific values of the tag that can be shared. All possible values if omitted.
   */
  readonly specificValues?: string[];
}

export interface SharedInternal extends BaseSharedTagProps {
  /**
   * A list of actions that can be performed on the tag.
   */
  readonly tagActions: TagAction[];
  /**
   * A list of actions on the tag with grant option, allowing grantees to further grant these permissions.
   */
  readonly tagActionsWithGrant?: TagAction[];
}

export interface SharedExternal extends BaseSharedTagProps {
  /**
   * A list of actions that can be performed on the tag. Only `TagAction.DESCRIBE` and `TagAction.ASSOCIATE` are allowed.
   */
  readonly tagActions: TagActionExternal[];
  /**
   * A list of actions on the tag with grant option, allowing grantees to further grant these permissions.
   */
  readonly tagActionsWithGrant?: TagActionExternal[];
}

export interface ShareProps {
  /**
   * Configurations for sharing LF-Tags with principals within the same AWS account.
   */
  readonly withinAccount?: SharedInternal[];
  /**
   * Configurations for sharing LF-Tags with external AWS accounts.
   */
  readonly withExternalAccount?: SharedExternal[];
}

export interface LFTagSharable extends LFTag {
  /**
   * OPTIONAL - Configuration detailing how the tag can be shared with specified principals.
   */
  readonly share?: ShareProps;
}
// ...
export interface LakePermission {
  /**
   * A list of principal identity ARNs (e.g., AWS accounts, IAM roles/users) that the permissions apply to.
   */
  readonly principals: string[];
  /**
   * LF tags associated with the permissions, used to specify fine-grained access controls.
   */
  readonly tags: LFTag[];
  /**
   * Actions that can be performed on databases, using Lake Formation Tag Based Access Control.
   */
  readonly databaseActions: DatabaseAction[];
  /**
   * OPTIONAL - Actions on databases with grant option, allowing grantees to further grant these permissions.
   */
  readonly databaseActionsWithGrant?: DatabaseAction[];
  /**
   * OPTIONAL - Actions that can be performed on tables, using Lake Formation Lake Formation Tag Based Access Control.
   */
  readonly tableActions?: TableAction[];
  /**
   * OPTIONAL - Actions on tables with grant option, allowing grantees to further grant these permissions.
   */
  readonly tableActionsWithGrant?: TableAction[];
}

export interface DlzLakeFormationAccessControlProps {
  /**
   * A list of Lake Formation tags that can be shared across accounts and principals.
   */
  readonly tags: LFTagSharable[];
  /**
   * A list of permission settings, specifying which Lake Formation permissions apply to which principals.
   */
  readonly permissions: LakePermission[];
}